Revert "move `generic-provider-views` and `Provider` to /plugins/Provider"
This reverts commit 66c8262ea2a575ffe06c2006b00317cea5f7f995.
