+++ src/generic-provider-views/Browser.js
