+++ src/generic-provider-views/Loader.js
